reorganizacao do repositorio para um package separado e correcao da poupanca

+++ src/conta/Conta.java
@@ -27,7 +27,7 @@ public class Conta {
 		this.numero = numero;
 	}
 
-	public void creditar(int valor) {
+	public void creditar(double valor) {
 		
 		this.saldo += valor;
 	}
